requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.7
Django==4.2.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
filelock==3.12.4
inflection==0.5.1
packaging==23.2
pipenv==2023.10.3
platformdirs==3.11.0
psycopg==3.1.12
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.24.5
